feat: contains peak filter
